moved viewmodeselector to UI
